refactor: move callbacks
